presentation

+++ mobile-app/SE1/app/src/main/java/co/edu/unal/se1/dataAccess/model/Account.java
@@ -11,7 +11,7 @@ public class Account implements Serializable{
     public int ID_Account;
 
     @ColumnInfo(name = "Balance")
-    public int Balance;
+    public double Balance;
 
     public int getID_Account() {
         return ID_Account;
@@ -21,11 +21,11 @@ public class Account implements Serializable{
         this.ID_Account = ID_Account;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return Balance;
     }
 
-    public void setBalance(int Balance) {
+    public void setBalance(double Balance) {
         this.Balance = Balance;
     }
 }
